update to add custom

+++ src/ImageBuilder.ts
@@ -49,6 +49,7 @@ export default class ImageBuilder {
     }
 
     async execute() {
+        imagebuilderRunStatus = ""
         try {
             azPath = await io.which("az", true);
             core.debug("Az module path: " + azPath);
@@ -390,12 +391,12 @@ export default class ImageBuilder {
 
     private async cleanup(subscriptionId: string) {
         try {
-            if (!this.isVhdDistribute && this.imgBuilderTemplateExists && this._taskParameters.actionRunMode == "full")) {
+            if (!this.isVhdDistribute && this.imgBuilderTemplateExists && this._taskParameters.actionRunMode == "full") {
                 await this._aibClient.deleteTemplate(this.templateName, subscriptionId);
                 console.log(`${this.templateName} got deleted`);
             }
 
-            if (storageAccountExists) {
+            if (storageAccountExists && this._taskParameters.actionRunMode != "nowait") {
                 let httpRequest: WebRequest = {
                     method: 'DELETE',
                     uri: this._client.getRequestUri(`subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Storage/storageAccounts/{storageAccount}`, { '{subscriptionId}': subscriptionId, '{resourceGroupName}': this._taskParameters.resourceGroupName, '{storageAccount}': this.storageAccount }, [], "2019-06-01")
